refactor: 엔티티 함수들의 폴더명 변경하기
