--- src/refactoring/utils/index.ts
@@ -1,4 +1,0 @@
-export * from './discountUtils';
-export * from './totalUtils';
-export * from './cartUtils';
-export * from './productUtils';
